requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
app==0.0.1
asgiref==3.4.1
async-timeout==4.0.2
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.72.0
fastapi-mail==1.0.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
packaging==21.3
protobuf==3.19.3
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-multipart==0.0.5
redis==4.0.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.17.0
wrapt==1.13.3
pandas~=1.4.0